third_party/blink/web_tests/inspector-protocol/emulation/device-emulation-partial-expected.txt

Tests that overriding only a single parameter does not affect others.

Running test: collectMetrics

Running test: noOverrides
[PASS]

Running test: width
[PASS]: width=300

Running test: height
[PASS]: height=400

Running test: deviceScaleFactor1
[PASS]: deviceScaleFactor=1

Running test: deviceScaleFactor2
[PASS]: deviceScaleFactor=2

Running test: clear
[PASS]

Running test: anotherWidth
[PASS]: width=400

Running test: anotherHeight
[PASS]: height=300

Running test: deviceScaleFactor3
[PASS]: deviceScaleFactor=3

Running test: clear
[PASS]
